chore: typecheck

+++ src/carboncopy/fs_utils.py
@@ -1,12 +1,14 @@
 import shutil
 from pathlib import Path
+from typing import Union, List, Dict, Any
 
 
-def clean_temp_files(path: Path):
-    shutil.rmtree(path, True)
+def clean_temp_files(path: Path) -> None:
+    if path:
+        shutil.rmtree(path, True)
 
 
-def get_template_file_paths(path):
+def get_template_file_paths(path: Path) -> List[Dict[str, Path]]:
     # TODO: expand to proper nested file structure.
     return [
         {"template": Path(filename), "destination": Path(filename).relative_to(path)}
@@ -14,7 +16,7 @@ def get_template_file_paths(path):
     ]
 
 
-def squash(source, destination):
+def squash(source: Path, destination: Path) -> None:
     try:
         shutil.copy(source, destination)
         print(
